Add options.noHighlight for hidden hints
Sometimes hints can be given, but they shouldn't highlight the whole
range.  A hint might have been hidden by the user, or we may want to
show an update card.  Our current case is an 'update date' card for a
plugin which maintains the date's state.  As per readme, you can now
supply an options object with the noHighlight option set to true. In
order to hide an item.

eg:

    EmberObject.create({
      info: {
        value: htmlValue, label: rdfaAnnotation.label,
        plainValue: userParsedDate,
        location: hint.location,
        hrId, hintsRegistry, editor
      },
      location: hint.location,
      card: this.get('who'),
      options: {
        noHighlight: true
      }
    })