Cleaning up code in components/rdfa-editor
This component passed some information to other component.  Most of
this is now only articulated in the templates.  We've also dropped a
bunch of this.get statements, as they are not needed anymore.  In case
methods are passed to objects which don't exist yet, a warning is
logged onto the console.