Optimisation of hints updating
An optimisation:
 a. decrease the amount of dead  DOMNodes when updating highlights
 b. potential performance gain.

With regard to a. the problem we try to 'tackle' is the fact there are
lots of hints which are unnecessary removed and highlighted by the
plugins. This results in unnecessary dom updates + the fact plugins
might work on 'dead' nodes. (Due to the whole async nature of the
processing and scheduling of the processing to the next runloop)

So now we only add and remove the real card delta cards. This helps
with the problem of dead nodes, but does not fix it