Lower highlight batch processing timeout
Our input keystrokes are now correctly batched.  Hence the batching of
highlights is mostly to ensure there's been time for users to interact
with our application again.  Lowering the timeout increases the speed
of visual feedback without (in contrast to earlier constructions) big
negative impacts on the user's appearance of typed characters.