Create stub export of ContextScanner
Plugins have based themselves on this component to import their
context-scanner.  They should be ported to use marawa instead or,
better yet, not be forced to execute any of this processing by
creating new objects.  In the current state, we're wrapping the
rdfa-context-scanner in a create interface so that other components
can keep using it as they used to.