Support non-task execute functions
Very simple applications can be ran in a synchronous loop.  There is
no need to have an async task in such a case, lowering the barrier to
entry into the framework.

This commit supports such native functions in the load-monitor.  The
detection logic might better live in the disptacher, that is not clear
yet.

Implementation could be cleaned using some filter logic, I was not
sure which one to use exactly so I'm adding it as it stands.