Add documentation for hint-card used in HintsRegistry#addHints
The addHints method of the HintsRegistry accepts an array of cards.
But what should these cards contain?  This commit introduces a new
method `addHint` so we can better document these hint cards.

It may be that the future will give us more explicit types.  In that
case the addHint method which we introduce here is an extra method to
maintain.  However, the bulk of the logic is split off from the
originating function.  The name of this function `_addHint` may need
to be revisited.