Further expanding the documentation
In discussions we have found the api to be wider than what was
originally documented.  These new options have been added to the
current documentation.  This documentation is still rough and each of
the implementations should be checked to verify they are complete.

Documentating a wide API like this one is not super friendly.  If this
gets updated more than once we should figure out a way to generate
this documentation from a stricture subset.  Such a change would make
this more maintainable.  The expanded format would still be needed so
tooling can operate on this information.