Add helper function to easen hints removal
Many plugins remove all of the hints for the region they are currently
scanning.  Doing this with the old API requires some basic logic, but
it's repeated everywhere.

This introduces a new method to easen the burden.  It removes the
hints for all the regions which an array of rdfaBlocks cover.  The
implementation itself is rather silly and could be improved over time,
yet plugins do become easier to read by it.