move service from addon to app and fix dynamic lookup
because of the way the config is imported this service needs to be in app.
We should probably consider better ways of getting the config as this feels a
bit hacky and might break in future versions of ember-cli/emberjs.

service lookup (of plugins) is now done using `getOwner(this).lookup`