Take nonbreaking spaces into account for visible strings.
Knowing the visible area of a string depends on a bunch of factors,
CSS included.  In this commit we altered the visible space of a string
to cope with nonbreaking spaces.  These tend to be almost always
visible.

The code did not make sense to me fully, there were three replacements
going on and I could not understand why the third replacement would
still be needed.  I'm also not sure about the first but couldn't find
a good test bed.  Let's assume it was added for good reason.

Also cleaned up is some indentation error.  It's important but I
didn't want to point it out with a separate commit :)