deprecate backSpace in favour of backspace
Documentation talks about backspace everywhere, but the method was
named backSpace.  For some reason that did not yield issues when
people were typing code but it's suboptimal to search for.  Deprecated
the original and replaced all use I could find by backspace.

I'm not sure this method really is a public method.  It is documented
as such but that may well be very incorrect.