Somewhat simple backspace-handler backspaceInEmptyNode
The functions currently use a whole bunch of if constructs, making it
hard to figure out what is really going on.  Trying to flatten those
so it's clearer what is going on.

The current version can backspace into the previous element which was
temporarily broken.  May well be due to my misunderstanding of the
code, I'm not sure.  Further refactoring is likely possible by
replacing the while loop alltogether as it only serves one case.