Simplify reading of backspaceInEmptyNode
The method backspaceInEmptyNode now uses recursion rather than a while
construct which is only used in one of its branches.  There are
however some strange issues with this code, stemming from my limited
understanding of the current code.  It seems removal is now working
incorrectly on a space before a hint.  Supposedly this stems from the
way whitespace is detected.

There are various approaches to solve that issue and we intend to
experiment in the hopes of finding a nice solution.