Prefer highlights to grow on the left side too
When highlights are received we should assume they grow on the left
side too so there's less chance for highlights not to be picked up.

Note that this approach has some downsides too.  Main issues arrise
from the fact that the cursor offset we use in the HintsRegistry (and
others) doesn't take the elements the user can see into account.
Untangling that situation is more complex, but would rectify a wider
set of issues.