changed cursor position handling in backspace handler
The general idea of the backspace handler now goes as follows:
- find the thing before the cursor.
- move the cursor before the thing using the browsers selection api
- try to remove that thing
- repeat until there is a visual difference
- update editor position with final caret position

We do this because the editor wants the cursor to be in a text node, but we
don't want to create new text nodes while backspacing. By depending on the
Selection api provided by the browser we can reposition the cursor without
needing text nodes. this has the benefit of working quite well with our change
detection loop.