try to be smarter when positioning cursor
when possible already place cursor at the end of the textNode before the actual
cursor position. This makes it easier to determine coordinates later on.

This is similar, but not exactly the same as what setCaret does. setCaret will
also consider the textNode after the provided position, which we explicitly
don't do here