Removed in lists/tab-handler: jumpToNextLiOfParentList
This is needlessly complicated. The way we make nested sublists (and
other editors too) is to insert it in the existing LI a new list,
and not, as I wrongly assumed a new LI with a list inside it.

So basically, when in a list, you just jump out of it, when in the
last li.