Remove removeWhitespaceNodes in list-helpers
When reconsidering this, it is not clear at all why it was
required in the first place.

Furthermore, it leads to unexpected behavior. You first grow a region
of adjacedent nodes. Then there is this 'removeWhitespaceNodes' on
this region (an array of nodes)  which
de facto removes all white spaces nodes. So you might end up in a
situation where the
remaining of your code expect it is an array of adjacedent nodes,
where in fact it is not.