Rename and change behaviour of growAdjacentNodesUntil
The function `growAdjacentNodesUntil` was not reflecting what it did:
region growing until a condition is met, starting from the provided node.
Hence the renaming to `growAdjacentRegionUntil`.
We still assume it is about adjacent nodes.

Furthermore: some removal of unrequired conditions, which should be
handled by the caller of the function, if deemded necessary.