Bugfix: editable lumpNode
In the case a lumpNode is the only node inside the editor, when
clicking on it, you could still edit it. This because editor.setCaret
couldn't deal with this case.

I choose to deal with the case myself, instead of fixing setCaret, as
we're moving away from it and don't want to open a box of pandora.

Current behaviour is simplistic, it will inject a textNode in case
```
<lumpNode></lumpNode><element></element>
```
Not suuper elegant, but better than before.

Since I expect ground works, I don't know if I want to refine this
more now.