Improve nbsp handling
strategy: when inserting a space or text that starts with spaces,
replace the space preceding the start of the range with an NBSP, if
there is one. Otherwise, insert as normal.

Adds charactersBefore and shiftedBy convenience methods to the
ModelPosition class.
Adds tests for the added methods.
Adds tests for the insert-text command