Fix lump node backspacing issues
The main problem was the contenteditable-false plugin triggering
at the same time as the lump node plugin when there was a lump node with
only a table.

That solved the backspace issues, but visually it was not right: the
cursor would "dissappear" when backspacing once, and backspacing again
would then trigger the normal lump node removal flow.

This was due to the rather strange way the browser renders a selection
that's anchored right after the table, and I think our css covering
that up. The fix was to make the lump node plugin aware of this
situation and handle it separately.