Needed support for @plugins in the debug component
Apparently, plugins use @plugins on the component to supply an array of
strings that indicate the plugins to be used. This has to be passed
through on the debug component as well. My mistake, was also not in the
README at all (maybe because that mechanism is fased out?). This
warrants a release of 0.50.0-beta.6.