Fix broken copying behavior
We were trying to create a range around an element on which we just
removed the parent. This gets detected as being the root node, so the
rangeAround method errors out since we cannot create a range around the
root.

Solved by using the gentreewalker, eliminating the need to create
the range in the first place.