Remove the embroider-optimized scenario
The Woodpecker server is being hit hard at the moment, running so many scenarios in parallel causes some of them to be killed due to out-of-memory issues.

As a temporary workaround we disable the embroider-optimized scenario since it fails anyway. In future PRs we can then replace the embroider-safe scenario with the more strict variant.