fix(rdfa): handle links to nodeless resources
Make it clear that it's usually not an error if a clickable backlink
can't go anywhere. This is an alternative to making the backlink
non-clickable automatically, but that has a performance impact since the
whole document is searched, and we'd be recalculating that on every
state change, for what should be a rare case anyway.