refactor(v2)!: make graphy work, fix some app build errors
v2 addons can't inject config into the app build, so the app needs
to explicitly import the webpack config we provide
as such the v2 conversion will necessarily be breaking

Additionally, fixes errors for @glint/template imports by using
import type { foo } instead of import { type foo }

They are apparently different. The former can be purely stripped by
babel because it is specced as never triggering side-effects.
The latter does trigger side-effects, so babel can't fully strip the
line, and it turns into import "@glint/template"

This in turn errors in the app build, because "@glint/template" isn't
a real package because it contains only types. You'd wish something in
the chain would be smarter about this, but it isn't.

This essentially means the latter syntax is to be avoided at all times.