Fix deletion of properties plugin to check the subject of links
Previously the plugin was checking backlinks from a deleted resource,
but was deleting any properties from the other end of the backlink, if
the predicate matched, so if there were multiple relationships, they
were all deleted. Now we check the subject of the property, to make sure
it matches the resource we're deleting.